requirements.txt
click>=8.1.0
google-generativeai>=0.8.0
tree-sitter>=0.21.0
tree-sitter-python>=0.21.0
tree-sitter-javascript>=0.21.0
jinja2>=3.1.0
python-dotenv>=1.0.0
pydantic>=2.0.0
flask>=3.0.0
